refactor code and impl redux
